refactor: rename emergency module
